Fixed 64-bit relative pointer calculation (int is 32 bit, long is 64 bit)
git-svn-id: https://svn.aros.org/svn/aros/trunk@34085 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/x86_64-all/include/aros/x86_64/cpu.h
@@ -84,7 +84,7 @@ do \
 {  \
     struct FullJumpVec *_v = v; \
     _v->jmp = 0xE9; \
-    _v->vec = (void *) ((unsigned int)(a) - (unsigned int)(_v) - 5); \
+    _v->vec = (void *) ((unsigned long)(a) - (unsigned long)(_v) - 5); \
 } while (0)
 
 /*
